requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
PyJWT==2.10.1
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3


#Flask==2.3.4
#Flask-SQLAlchemy==3.0.3
#Flask-Cors==4.0.1
#python-dotenv==1.0.0
#PyJWT==2.8.0
#passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.7
